Fix new errors introduced when restructuring files structure.
